Add destroy functionality, sad paths in application controller. BDD, not tested
