+++ config/routes.rb
@@ -1,4 +1,8 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  resources :books, only: %i[index create destroy]
+  namespace :api do
+    namespace :v1 do
+      resources :books, only: %i[index create destroy]
+    end
+  end
 end
